LPS-75598 user-associated-data-api - moves classes to the appropriate packages
