+++ modules/apps/foundation/user-associated-data/user-associated-data-api/bnd.bnd
@@ -2,5 +2,8 @@ Bundle-Name: Liferay User Associated Data API
 Bundle-SymbolicName: com.liferay.user.associated.data.api
 Bundle-Version: 1.0.0
 Export-Package:\
-	com.liferay.user.associated.data.model,\
+	com.liferay.user.associated.data.aggregator,\
+	com.liferay.user.associated.data.anonymizer,\
+	com.liferay.user.associated.data.entity,\
+	com.liferay.user.associated.data.exporter,\
 	com.liferay.user.associated.data.registry
