+++ modules/apps/foundation/user-associated-data/user-associated-data-api/src/main/java/com/liferay/user/associated/data/entity/UADEntity.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.user.associated.data.model;
+package com.liferay.user.associated.data.entity;
 
 /**
  * @author William Newbury
