+++ modules/apps/foundation/user-associated-data/user-associated-data-api/src/main/java/com/liferay/user/associated/data/registry/UADRegistryUtil.java
@@ -14,9 +14,9 @@
 
 package com.liferay.user.associated.data.registry;
 
-import com.liferay.user.associated.data.model.UADEntityAggregator;
-import com.liferay.user.associated.data.model.UADEntityAnonymizer;
-import com.liferay.user.associated.data.model.UADEntityExporter;
+import com.liferay.user.associated.data.aggregator.UADEntityAggregator;
+import com.liferay.user.associated.data.anonymizer.UADEntityAnonymizer;
+import com.liferay.user.associated.data.exporter.UADEntityExporter;
 
 import java.util.Collection;
