+++ modules/apps/foundation/user-associated-data/user-associated-data-api/src/main/resources/com/liferay/user/associated/data/anonymizer/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
